# api-backend/openapi.yaml
openapi: 3.0.0
info:
  title: MeusFilmes API
  version: 1.0.0
  description: API for the MeusFilmes application, providing access to movie data, user lists, and watched history.
servers:
  - url: /api # Adjusted to reflect Nginx proxy
    description: Main API access via Nginx proxy
paths:
...
  /movie/watched/:
    post:
      summary: Marca um filme como assistido para o usuário autenticado
      tags:
        - Watched
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tmdb_id:
                  type: integer
                  example: 603692
                watched_at:
                  type: string
                  format: date
                  example: '2024-05-20'
              required:
                - tmdb_id
                - watched_at
      responses:
        '201':
          description: Filme marcado como assistido
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  tmdb_id:
                    type: integer
                  watched_at:
                    type: string
                  msg:
                    type: string
        '200':
          description: Data de assistido atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  tmdb_id:
                    type: integer
                  watched_at:
                    type: string
                  msg:
                    type: string
        '400':
          description: Dados ausentes ou inválidos
        '500':
          description: Erro ao salvar filme assistido
    get:
      summary: Lista todos os filmes assistidos do usuário autenticado
      tags:
        - Watched
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de filmes assistidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  watched:
                    type: array
                    items:
                      type: object
                      properties:
                        tmdb_id:
                          type: integer
                        watched_at:
                          type: string
  /movie/watched/{tmdb_id}:
    delete:
      summary: Remove um filme da lista de assistidos do usuário autenticado
      tags:
        - Watched
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: tmdb_id
          schema:
            type: integer
          required: true
          description: ID do filme no TMDB
      responses:
        '200':
          description: Filme removido da lista de assistidos
        '404':
          description: Filme não encontrado na lista de assistidos
  /movie/watched/count:
    get:
      summary: Conta o número de filmes assistidos pelo usuário autenticado, com filtro opcional por data
      tags:
        - Watched
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: start_date
          schema:
            type: string
            format: date
          required: false
          description: Data inicial (YYYY-MM-DD)
        - in: query
          name: end_date
          schema:
            type: string
            format: date
          required: false
          description: Data final (YYYY-MM-DD)
      responses:
        '200':
          description: Contagem de filmes assistidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: integer
                  count:
                    type: integer
        '400':
          description: Formato de data inválido
  /tmdb/config:
    get:
      summary: Obtém a configuração do TMDB (chave de API, base de imagens, etc)
      tags:
        - TMDB
      responses:
        '200':
          description: Configuração do TMDB retornada com sucesso
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar configuração do TMDB
  /tmdb/popular:
    get:
      summary: Lista filmes populares do TMDB
      tags:
        - TMDB
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Página dos resultados
      responses:
        '200':
          description: Lista de filmes populares
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar filmes populares
  /tmdb/movie/{tmdb_id}:
    get:
      summary: Detalhes de um filme pelo TMDB ID (usa cache local se disponível)
      tags:
        - TMDB
      parameters:
        - in: path
          name: tmdb_id
          schema:
            type: integer
          required: true
          description: ID do filme no TMDB
      responses:
        '200':
          description: Detalhes do filme
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Filme não encontrado
        '500':
          description: Erro inesperado
  /tmdb/search:
    get:
      summary: Busca filmes no TMDB por texto
      tags:
        - TMDB
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Texto de busca
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Página dos resultados
      responses:
        '200':
          description: Resultados da busca
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar filmes
  /tmdb/movie/{tmdb_id}/credits:
    get:
      summary: Créditos (elenco e equipe) de um filme
      tags:
        - TMDB
      parameters:
        - in: path
          name: tmdb_id
          schema:
            type: integer
          required: true
          description: ID do filme no TMDB
      responses:
        '200':
          description: Créditos do filme
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar créditos
  /tmdb/movie/{tmdb_id}/recommendations:
    get:
      summary: Recomendações de filmes similares
      tags:
        - TMDB
      parameters:
        - in: path
          name: tmdb_id
          schema:
            type: integer
          required: true
          description: ID do filme no TMDB
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Página dos resultados
      responses:
        '200':
          description: Lista de recomendações
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar recomendações
  /tmdb/genres:
    get:
      summary: Lista de gêneros de filmes (legado)
      tags:
        - TMDB
      responses:
        '200':
          description: Lista de gêneros
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar gêneros
  /tmdb/movie/genres:
    get:
      summary: Lista de gêneros de filmes (novo endpoint)
      tags:
        - TMDB
      responses:
        '200':
          description: Lista de gêneros
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Erro ao buscar gêneros
  /tmdb/discover/movie:
    get:
      summary: Descobre filmes por gênero
      tags:
        - TMDB
      parameters:
        - in: query
          name: with_genres
          schema:
            type: string
          required: true
          description: IDs dos gêneros separados por vírgula
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Página dos resultados
      responses:
        '200':
          description: Lista de filmes descobertos
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Parâmetro de gênero ausente
        '500':
          description: Erro ao buscar filmes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []

# You will need to manually add definitions for all your other API endpoints
# (e.g., /login, /lists, /movie/watched, /tmdb/*, etc.)
# following the OpenAPI 3.0 specification format.
# Pay close attention to request bodies, parameters (path, query), and responses for each.